Rename pdffile modules to pdf_file
