+++ app/routes/pdf_files.py
@@ -3,8 +3,8 @@ from fastapi import APIRouter, Depends, UploadFile, File, Form, HTTPException
 from sqlalchemy.orm import Session
 from fastapi.responses import FileResponse
 from app.dependencies import get_db
-from app.schemas.pdffile import PDFFileCreate, PDFFileResponse
-from app.crud import pdffile as crud_pdffile
+from app.schemas.pdf_file import PDFFileCreate, PDFFileResponse
+from app.crud import pdf_file as crud_pdf_file
 import os
 
 router = APIRouter(prefix="/pdf", tags=["PDF"])
@@ -12,7 +12,7 @@ router = APIRouter(prefix="/pdf", tags=["PDF"])
 
 @router.get("/", response_model=List[PDFFileResponse])
 def list_pdfs(db: Session = Depends(get_db)):
-    return crud_pdffile.get_multi(db)
+    return crud_pdf_file.get_multi(db)
 
 
 @router.post("/", response_model=PDFFileResponse, status_code=201)
@@ -23,14 +23,14 @@ async def upload_pdf(
 ):
     if file.content_type != "application/pdf":
         raise HTTPException(400, "Il file deve essere un PDF")
-    return crud_pdffile.create(db, obj_in=PDFFileCreate(title=title), file=file)
+    return crud_pdf_file.create(db, obj_in=PDFFileCreate(title=title), file=file)
 
 
 @router.get("/{filename}")
 def get_pdf(filename: str):
     """Return a previously uploaded PDF by filename."""
     from pathlib import Path
-    from app.crud.pdffile import get_upload_root
+    from app.crud.pdf_file import get_upload_root
 
     # Prevent path traversal attacks by stripping directory components
     safe_name = Path(filename).name
